Renamed main teleOp to Competition TeleOP
